requirements.txt
black==25.11.0
click==8.3.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
python-dotenv==1.2.1
pytokens==0.3.0
ruff==0.14.7
